Menambahkan validasi body dan query

+++ server.js
@@ -11,13 +11,23 @@ const pool = new Pool({
 	password: 'e',
 });
 
-// function isValid(req, requiredAttribute) {
-//   let valid;
-//   requiredAttribute.every((val) => {
-//     valid = ((req.query[val] !== undefined) && (req.query[val] !== ''));
-//     return valid;
-//   });
-// }
+function isQueryStringValid(req, requiredAttribute) {
+  let valid;
+  requiredAttribute.every((val) => {
+    valid = ((req.query[val] !== undefined) && (req.query[val] !== ''));
+    return valid;
+  });
+  return valid;
+}
+
+function isBodyValid(req, requiredAttribute) {
+  let valid;
+  requiredAttribute.every((val) => {
+    valid = ((req.body[val] !== undefined) && (req.body[val] !== ''));
+    return valid;
+  });
+  return valid;
+}
 
 express()
   .use(bodyParser.json())
@@ -34,17 +44,21 @@ express()
     next();
   })
 	.post('/pengantri', (req, res) => {
-    pool.connect().then((client) => {
-      return client.query('INSERT INTO pengantri (namawakil, jumlah, waktumulai) SELECT $1, $2, (current_time at time zone 'cxt')::TIME(0);', [req.body.namawakil, req.body.jumlah])
-        .then(() => {
-          client.release();
-          res.status(200).send('Pengantri berhasil ditambahkan');
-        }).catch((err) => {
-          client.release();
-          console.log(err.stack);
-          res.status(500).send('Pengantri gagal ditambahkan');
-        })
-    })
+    if (isBodyValid(req, ['namawakil', 'jumlah'])) {
+      pool.connect().then((client) => {
+        return client.query("INSERT INTO pengantri (namawakil, jumlah, waktumulai) SELECT $1, $2, (current_time at time zone 'cxt')::TIME(0);", [req.body.namawakil, req.body.jumlah])
+          .then(() => {
+            client.release();
+            res.status(200).send('Pengantri berhasil ditambahkan');
+          }).catch((err) => {
+            client.release();
+            console.log(err.stack);
+            res.status(500).send('Pengantri gagal ditambahkan');
+          })
+      })
+    } else {
+      res.status(400).send("Request body tidak lengkap");
+    }
 	})
   .get('/pengantri', (req, res) => {
     pool.connect().then((client) => {
@@ -61,29 +75,33 @@ express()
       })
 	})
   .delete('/pengantri', (req, res) => {
-    pool.connect().then((client) => {
-      return client.query('DELETE FROM pengantri WHERE id=$1;', [req.query.id]).then(() => {
-        client.query('SELECT COUNT(*) FROM pengantri;').then((rescount) => {
-          if (rescount.rows[0].count === '0') {
-            client.query('ALTER SEQUENCE pengantri_id_seq RESTART WITH 1;').then(() => {
+    if (isQueryStringValid(req, ['id'])) {
+      pool.connect().then((client) => {
+        return client.query('DELETE FROM pengantri WHERE id=$1;', [req.query.id]).then(() => {
+          client.query('SELECT COUNT(*) FROM pengantri;').then((rescount) => {
+            if (rescount.rows[0].count === '0') {
+              client.query('ALTER SEQUENCE pengantri_id_seq RESTART WITH 1;').then(() => {
+                client.release();
+                res.status(200).send(`Pengantri ${req.query.id} berhasil dihapus`);
+              }).catch(err => {
+                throw err;
+              });
+            } else {
               client.release();
               res.status(200).send(`Pengantri ${req.query.id} berhasil dihapus`);
-            }).catch(err => {
-              throw err;
-            });
-          } else {
-            client.release();
-            res.status(200).send(`Pengantri ${req.query.id} berhasil dihapus`);
-          }
+            }
+          }).catch(err => {
+            throw err;
+          });
         }).catch(err => {
-          throw err;
-        });
-      }).catch(err => {
-        client.release();
-        console.log(err.stack);
-        res.status(500).send(`Pengantri ${req.query.id} gagal dihapus`);
+          client.release();
+          console.log(err.stack);
+          res.status(500).send(`Pengantri ${req.query.id} gagal dihapus`);
+        })
       })
-    })
+    } else {
+      res.status(400).send("Request body tidak lengkap");
+    }
   })
 	.get('/meja', (req, res) => {
     pool.connect().then((client) => {
@@ -98,39 +116,30 @@ express()
         })
     })
 	})
-  .put('/meja/kosong', (req, res) => {
-    pool.connect().then((client) => {
-      return client.query('UPDATE meja SET status = false WHERE id = $2 AND kapasitas = $3', [req.query.id, req.query.kapasitas])
-        .then(() => {
-          client.release();
-          res.status(200).send(`Status meja ${req.query.id} berhasil diganti`);
-        }).catch((err) => {
-          client.release()
-          console.log(err.stack);
-          res.status(500).send('Status meja gagal diganti');
-        })
-    })
-  })
   .put('/meja', (req, res) => {
-    let query = `UPDATE meja SET status = ${req.body.status} WHERE `;
-    if (req.body.meja.length === 0) {
-      res.status(200).send('Tidak ada meja yang statusnya diubah');
-    } else {
-      query += `(id = ${req.body.meja[0].id} AND kapasitas = ${req.body.meja[0].kapasitas})`
-      for (let i=1; i<=req.body.meja.length-1; i++) {
-        query += ` OR (id = ${req.body.meja[i].id} AND kapasitas = ${req.body.meja[i].kapasitas})`
+    if (isBodyValid(req, ['status', 'meja'])) {
+      let query = `UPDATE meja SET status = ${req.body.status} WHERE `;
+      if (req.body.meja.length === 0) {
+        res.status(200).send('Tidak ada meja yang statusnya diubah');
+      } else {
+        query += `(id = ${req.body.meja[0].id} AND kapasitas = ${req.body.meja[0].kapasitas})`
+        for (let i=1; i<=req.body.meja.length-1; i++) {
+          query += ` OR (id = ${req.body.meja[i].id} AND kapasitas = ${req.body.meja[i].kapasitas})`
+        }
       }
+      pool.connect().then((client) => {
+        return client.query(query)
+          .then(() => {
+            client.release();
+            res.status(200).send(`Status meja berhasil diganti`);
+          }).catch((err) => {
+            client.release()
+            console.log(err.stack);
+            res.status(500).send('Status meja gagal diganti');
+          })
+      })
+    } else {
+      res.status(400).send("Request body tidak lengkap");
     }
-    pool.connect().then((client) => {
-      return client.query(query)
-        .then(() => {
-          client.release();
-          res.status(200).send(`Status meja berhasil diganti`);
-        }).catch((err) => {
-          client.release()
-          console.log(err.stack);
-          res.status(500).send('Status meja gagal diganti');
-        })
-    })
   })
 	.listen(PORT, () => console.log(`Listening on ${PORT}`));
